move visual loader
