+++ src/factory/utils/index.js
@@ -1,2 +1,3 @@
 export { default as aggregateData } from './aggregateData';
 export { default as assignQueryAlias } from './assignQueryAlias';
+export { default as visualLoader } from './visualLoader';
